Modularized turtle movements into a class
