+++ scripts/ros_works_turtlesim​/turtle_closed_loop.py
@@ -1,11 +1,14 @@
 #!/usr/bin/env python
+# coding: utf-8
 
 import rospy
 from geometry_msgs.msg import Twist
 from turtlesim.msg import Pose
 import math
 
-class TurtleClosedLoop:
+from turtle_kinematics import TurtleKinematics
+
+class TurtleClosedLoop(TurtleKinematics):
     ''' Classe para o controle de malha fechada da tartaruga '''
     def __init__(self):
         ''' Construtor da classe TurtleClosedLoop '''
@@ -26,40 +29,13 @@ class TurtleClosedLoop:
         self.turtle_angular_velocity = 0
 
     def callback(self, msg):
-        ''' Funcao chamada toda a vez que atualizar-se o topico: turtle1/pose '''
+        ''' Funcao chamada toda a vez que atualiza-se o topico: turtle1/pose '''
         self.turtle_x = msg.x
         self.turtle_y = msg.y
         self.turtle_theta = msg.theta
         self.turtle_linear_velocity = msg.linear_velocity
         self.turtle_angular_velocity = msg.angular_velocity
         # print 'x = ',self.turtle_x,'\ny = ',self.turtle_y
-
-    def move_angular(self, vel):
-        ''' Cria e envia mensagem de velocidade angular '''
-        vel_msg = Twist()
-        vel_msg.angular.z = vel
-        self.pub.publish(vel_msg)
-        turtle.rate.sleep()
-
-    def move_linear(self, vel):
-        ''' Cria e envia mensagem de velocidade linear '''
-        vel_msg = Twist()
-        vel_msg.linear.x = vel
-        self.pub.publish(vel_msg)
-        turtle.rate.sleep()
-
-    def move_general(self, vel, ang):
-        ''' Recebe velocidades linear e angular para movimentar a tartaruga '''
-        vel_msg = Twist()
-        vel_msg.linear.x = vel
-        vel_msg.angular.z = ang
-        self.pub.publish(vel_msg)
-        turtle.rate.sleep()
-
-    def stop(self):
-        ''' Cria e publica mensagem para parar a tartaruga '''
-        vel_msg = Twist()
-        self.pub.publish(vel_msg)
 
     def go_to_point(self, x, y):
         '''
@@ -95,6 +71,7 @@ class TurtleClosedLoop:
 
 
 def interface_usuario():
+    ''' Interface de Usuario Principal '''
     print 'Aonde deseja ir?'
     print 'X: '
     x = input()
